clean-up examples folder
