--- examples/utils/download_extract.py
@@ -1,15 +1,0 @@
-import argparse
-import logging
-
-from torchtext.utils import download_from_url, extract_archive
-
-parser = argparse.ArgumentParser(description="Download and extract a given dataset")
-parser.add_argument("--url", default="http://www.quest.dcs.shef.ac.uk/wmt16_files_mmt/" "validation.tar.gz")
-parser.add_argument("--data", default="./validation.tar.gz")
-parser.add_argument("--logging-level", default="WARNING")
-args = parser.parse_args()
-
-logging.basicConfig(level=getattr(logging, args.logging_level))
-
-tar_file = download_from_url(args.url, args.data)
-extracted_files = extract_archive(args.data, "extracted_files")
